belajar data di python

+++ CobaVar.java
@@ -5,6 +5,6 @@ public class CobaVar {
         double tinggiBadan = 153;
         System.out.println("Nama Saya " + namaDepan);
         System.out.println("Umur saya " + umur);
-        System.out.println("Tinggi badan saya " + tinggiBadan);
+        System.out.println("Tinggi badan saya " + (int) tinggiBadan);
     }    
 }
